+++ src/wfcomman.c
@@ -1180,7 +1180,7 @@ AppCommandProc(register DWORD id)
 	  if(szFiles == NULL && pDataObj != NULL && pDataObj->lpVtbl->GetData(pDataObj, &fmtetcDrop, &stgmed) == S_OK)
 	  {
 	  	LPWSTR lpFile = GlobalLock(stgmed.hGlobal);
-	  	DWORD cchFile = wcslen(lpFile);
+	  	SIZE_T cchFile = wcslen(lpFile);
 		szFiles = (LPWSTR)LocalAlloc(LMEM_FIXED, (cchFile+3) * sizeof(WCHAR));
 		lstrcpy (szFiles+1, lpFile);
 		*szFiles = '\"';
@@ -1638,7 +1638,7 @@ AppCommandProc(register DWORD id)
 
       } else {
          dwSuperDlgMode = id;
-         ret = DialogBox(hAppInstance, MAKEINTRESOURCE(CHOOSEDRIVEDLG), hwndFrame, ChooseDriveDlgProc);
+         ret = (INT)DialogBox(hAppInstance, MAKEINTRESOURCE(CHOOSEDRIVEDLG), hwndFrame, ChooseDriveDlgProc);
       }
 
       break;
@@ -1971,7 +1971,7 @@ ChangeDisplay:
 
        // toggle pluses view bit
 
-       dwFlags = GetWindowLongPtr(hwndActive, GWL_VIEW) ^ VIEW_PLUSES;
+       dwFlags = (DWORD)(GetWindowLongPtr(hwndActive, GWL_VIEW) ^ VIEW_PLUSES);
 
        SetWindowLongPtr(hwndActive, GWL_VIEW, dwFlags);
